+++ src/main/java/com/crud/demo/product/CruddemoApplication.java
@@ -1,4 +1,4 @@
-package com.crud.demo;
+package com.crud.demo.product;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
