+++ src/main/java/com/crud/demo/product/feign/facade/ProductFacade.java
@@ -1,13 +1,13 @@
-package com.crud.demo.feign.facade;
+package com.crud.demo.product.feign.facade;
 
 import java.util.List;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
 
-import com.crud.demo.dao.ProductDao;
-import com.crud.demo.dao.ProductsDao;
-import com.crud.demo.feign.client.ProductIF;
+import com.crud.demo.product.dao.ProductDao;
+import com.crud.demo.product.dao.ProductsDao;
+import com.crud.demo.product.feign.client.ProductIF;
 
 @Component
 public class ProductFacade {
