+++ src/main/java/com/crud/demo/product/service/ProductServiceIF.java
@@ -1,0 +1,12 @@
+package com.crud.demo.product.service;
+
+import java.util.List;
+
+import com.crud.demo.product.dao.ProductDao;
+import com.crud.demo.product.dao.ProductsDao;
+
+public interface ProductServiceIF {
+
+	ProductsDao getProduct();
+
+}
