--- src/main/java/com/crud/demo/service/ProductServiceIF.java
@@ -1,12 +1,0 @@
-package com.crud.demo.service;
-
-import java.util.List;
-
-import com.crud.demo.dao.ProductDao;
-import com.crud.demo.dao.ProductsDao;
-
-public interface ProductServiceIF {
-
-	ProductsDao getProduct();
-
-}
